fix: bugs in authorization

+++ rsclone/src/js/core/services/request.js
@@ -16,20 +16,20 @@ export class Request {
     };
     await fetch(Constants.PathSignUp, requestOptions)
       .then((response) => {
-        if (response.status >= 400 && response.status < 600) {
+        if (response.status >= 400 && response.status < 500) {
           message.innerHTML = 'Fill in the fields correctly.Try again';
-          setTimeout(() => {
-            message.innerHTML = 'Sign in';
-          }, 2000);
-        } else {
-          message.innerHTML = 'Registration was successful';
+          setTimeout(() => { message.innerHTML = 'Sing Up'; }, 2000);
+        } else if (response.status >= 500 && response.status < 600) {
+          this.errorMessage('Sorry, the server is temporarily unresponsive. Try later');
+        } else if (response.status >= 200 && response.status < 300) {
+          this.errorMessage('Registration was successful');
           setTimeout(() => {
             finish();
           }, 1000);
         }
         return response.json();
       }).then((result) => this.setUserDataInStorage(result))
-      .catch((err) => console.err(err));
+      .catch((err) => console.log(err));
   }
 
   // Войти POST
@@ -43,20 +43,22 @@ export class Request {
     };
     await fetch(Constants.PathSignIn, requestOptions)
       .then((response) => {
-        if (response.status >= 400 && response.status < 600) {
+        if (response.status >= 400 && response.status < 500) {
+          setTimeout(() => { message.innerHTML = 'Sing In'; }, 2000);
           message.innerHTML = 'Fill in the fields correctly.Try again';
-          setTimeout(() => {
-            message.innerHTML = 'Authorization';
-          }, 2000);
-        } else {
-          message.innerHTML = 'Authorization was successful';
+        } else if (response.status >= 500 && response.status < 600) {
+          this.errorMessage('Sorry, the server is temporarily unresponsive. Try later');
+        } else if (response.status >= 200 && response.status < 300) {
+          this.errorMessage('Authorization was successful');
           setTimeout(() => {
             finish();
           }, 1000);
         }
         return response.json();
-      }).then((result) => this.setUserDataInStorage(result))
-      .catch((err) => { console.err(err); });
+      }).then((result) => {
+        this.setUserDataInStorage(result);
+      })
+      .catch((err) => { console.log(err); });
   }
 
   async logout() {
@@ -69,16 +71,19 @@ export class Request {
     };
     await fetch(Constants.PathLogout, requestOptions)
       .then((response) => {
-        if (response.status >= 400 && response.status < 600) {
+        if (response.status >= 400 && response.status < 500) {
           throw new Error(`Wrong status: ${response.status}`);
-        } else if (response.status === 200) {
+        } else if (response.status >= 500 && response.status < 600) {
+          this.errorMessage('Sorry, the server is temporarily unresponsive. Try later');
+        } else if (response.status >= 200 && response.status < 300) {
           localStorage.clear();
           Storage.setAuthorizationStatus('false');
           Storage.setUserStatisticData(Constants.EmptyUserData);
+          Storage.setUserName('Anonymous');
           removeUserAuthorizationData();
         }
       })
-      .catch((err) => console.err(err));
+      .catch((err) => console.log(err));
   }
 
   // Получить данные клиента GET
@@ -91,9 +96,14 @@ export class Request {
       },
     };
     await fetch(Constants.PathGetAndPutRequest, requestOptions)
-      .then((response) => response.json())
+      .then((response) => {
+        if (response.status >= 500 && response.status < 600) {
+          this.errorMessage('Sorry, the server is temporarily unresponsive. Try later');
+        }
+        return response.json();
+      })
       .then((result) => Storage.setUserStatisticData(result.data))
-      .catch((err) => console.err(err));
+      .catch((err) => console.log(err));
   }
 
   // Получить данные статистики PUT
@@ -107,8 +117,13 @@ export class Request {
       body: JSON.stringify(data),
     };
     await fetch(Constants.PathGetAndPutRequest, requestOptions)
-      .then((response) => response.json())
-      .catch((err) => console.err(err));
+      .then((response) => {
+        if (response.status >= 500 && response.status < 600) {
+          this.errorMessage('Sorry, the server is temporarily unresponsive. Try later');
+        }
+        return response.json();
+      })
+      .catch((err) => console.log(err));
   }
 
   setUserDataInStorage(result) {
@@ -119,5 +134,15 @@ export class Request {
 
   setTokenInStorage(token) {
     Storage.setUserToken(token);
+  }
+
+  errorMessage(message) {
+    const messageBlock = document.createElement('div');
+    messageBlock.classList.add('error-message');
+    messageBlock.innerHTML = message;
+    document.querySelector('body').append(messageBlock);
+    setTimeout(() => {
+      document.querySelector('body').removeChild(messageBlock);
+    }, 2000);
   }
 }
